% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blast.R
\name{filter_blast}
\alias{filter_blast}
\title{Simple BLAST-results filter}
\usage{
filter_blast(raw_granges, min_len = 0, max_e = 100, min_identity = 0)
}
\arguments{
\item{raw_granges}{A \linkS4class{GRangesList} as
returned from \link{read_blast_xml}.}

\item{min_len, max_e, min_identity}{Integer or float to filter the
BLAST-results by minimum length, maximum e-value or minimum percent idenity}

\item{grange_output}{If set TRUE, output it a
\linkS4class{GRangesList} object, otherwise a
\link[base]{data.frame} object.}

\item{keep_sequences}{If FALSE, sequence and alignment columns are discarded.}
}
\description{
Filter BLAST-results by different criteria. Removes every
alignment not fulfilling the criteria
}
