% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIS_functions.R
\name{get_best_start}
\alias{get_best_start}
\title{Get best start for WIS}
\usage{
get_best_start(granges_list, overhang = 0)
}
\arguments{
\item{granges_list}{A \linkS4class{GRangesList} object
containing the intervals.}

\item{overhang}{A non-negative integer specifying for a linearized circular
scale how much of the end is overlapping with the start.}
}
\value{
Returns the index of the optimal interval to start a WIS-algorithm with. If no
such interval exists, returns an integer vector with indices for the best
possible interval and all intervalls overlapping its starting position.
}
\description{
For a circular scale, this function finds the best interval(s)
to start the WIS alsogrithm with.
}
\details{
Takes a circular  \linkS4class{GRangesList}
object and finds the best intervals to start with. Ideally, this is an
inverval, whose start position is not overlapped by any other interval. If
no such interval exists, the interval with the least overlaps at its start
position and each interval overlapping this position are returned.

Strand information is ignored here, has to be taken into account before
calling this fucntion.

Also checks if overhang is long enough or should be extended, e.g. in
BLAST-searches by copying a longer sequence from the beginning to the end.
}
