% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIS_functions.R
\name{log_factorial}
\alias{log_factorial}
\title{Log factorial}
\usage{
log_factorial(n)
}
\arguments{
\item{n}{Integer or vector of integers}
}
\value{
log(n!) as float. If input is a vector of integers, output is a
vecor with log(n!) for each value in n
}
\description{
Calculates the natural logarithm of a factorial for high numbers
}
