% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blast.R
\name{read_blast_xml}
\alias{read_blast_xml}
\title{Parse BLAST XML output}
\usage{
read_blast_xml(x, grange_output = T, keep_sequences = F)
}
\arguments{
\item{x}{A xml file created with NCBI-BLAST option -outfmt 5 or AB-BLAST
option -mformat=7.}

\item{grange_output}{If set TRUE, output is a \linkS4class{GRangesList}
object, otherwise a \link[base]{data.frame} object.}

\item{keep_sequences}{If FALSE, sequence and alignment columns are discarded.}
}
\value{
A list with two elements accessible via: \cr \code{$metadata}:
BLAST-metadata such as run parameters, sequence information and
statistics.\cr \code{$alignments}: Information about the alignments.\cr If
\code{grange_output} is set FALSE, output is a \link[base]{data.frame}
object, else a \linkS4class{GRangesList} object with start, end, strand,
sequence name and length taken from the subject. Information about the
query is presered in the metadata columns.
}
\description{
Parses a BLAST xml-file to a
\linkS4class{GRangesList} or \link[base]{data.frame} object.
}
\details{
This function parses BLAST-xml output files using
\link[XML]{xmlParse}. Setting \code{keep_sequences} FALSE will discard all
sequence and alignment columns to save memory.
}
