% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIS_functions.R
\name{run_get_wis}
\alias{run_get_wis}
\title{Run get_wis function}
\usage{
run_get_wis(
  granges_list,
  is_circular,
  overhang = 0,
  use_strand = T,
  frame_col = NULL,
  score_col = "score",
  score_f = function(x, y, n) x + y,
  max_f = function(x, y, n) x + y,
  overlap = 0
)
}
\arguments{
\item{granges_list}{A set of intervals as a
\linkS4class{GRangesList} object. Requires a
interval-score to maximize over as a column specified in \code{score_col}.}

\item{is_circular}{Boolean, set TRUE if scale is circular.}

\item{overhang}{A non-negative integer. A non-negative integer specifying for
a linearized circular scale how much of the end is overlapping with the
start. Ignored if is_circular is FALSE.}

\item{use_strand}{Boolean, if TRUE, overlapping intervals on different strand
are considered compatible.}

\item{frame_col}{Optional, column in \code{granges_list} containing frame information. If
set, overlapping intervals on different strand are considered compatible.}

\item{score_col}{Name of a column in \code{granges_list} containing integers or
floats to be used as a score to maximize over.}

\item{score_f}{Function to update the score of a set each time an element is
added. Passed to \code{score_funct} in \link{get_wis}, see there for detailed
description.}

\item{max_f}{Function to calculate a set-score to maximize over. Passed to
\code{max_funct} in \link{get_wis}, see there for detailed description.}

\item{overlap}{A non-negative integer. The allowed overlap between two
intervals.}
}
\value{
A \linkS4class{GRangesList} object with intervals
compatible according to the specified options.
}
\description{
Runs the \link{get_wis} function on a set of weighted intervals given
as a \linkS4class{GRangesList} object by taking into
account several parameters.
}
\details{
This function runs the \link{get_wis} function on a set of intervals and allows to distinguish between intervals on different stradns or frames. For each of these settings,\link{get_wis} is run seperately. For detailed description of the output, see \link{get_wis}.
}
