% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIS_functions.R
\name{sum_score}
\alias{sum_score}
\title{BLAST sum-score}
\usage{
sum_score(b, metadata)
}
\arguments{
\item{b}{Vector of integers or floats.}

\item{metadata}{BLAST metadata as returned by \link{read_blast_xml}.
A \link[base]{data.frame} object with one row containing the columns
Statistics_lambda, Statistics_kappa, query_len and Statistics_db.le}
}
\value{
Sum score as float.
}
\description{
Caclulated BLAST sum-score from raw-scores and BLAST search
statistics
}
\details{
This function calculates the BLAST sum-score as described  in
Equation 4-17 here:
\url{http://etutorials.org/Misc/blast/Part+II+Theory/Chapter+4.+Sequence+Similarity/4.7+Sum+Statistics+and+Sum+Scores/}
}
